Move assets of Races to client dir
