+++ app/views/shared/_races.html.haml
@@ -1,1 +1,1 @@
-= react_component 'Races', url: '/api/races.json'
+= react_component('Races', props: {url: '/api/races.json'}, prerender: false)
