+++ client/app/bundles/components/Race.jsx
@@ -1,44 +1,12 @@
-class Races extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {data: []};
-  }
+import React, { PropTypes } from 'react';
+import RaceOwner from './RaceOwner';
+import Candidates from './Candidates';
 
-  componentDidMount() {
-    $.ajax({
-      url: this.props.url,
-      dataType: 'json',
-      success: function(result) {
-        this.setState({data: result});
-      }.bind(this),
-      error: function(xhr, status, err) {
-        console.error(this.props.url, status, err.toString());
-      }.bind(this)
-    });
-  }
+export default class Race extends React.Component {
+  static propTypes = {
+    data: React.PropTypes.object.isRequired,
+  };
 
-  renderRace(race) {
-    return <Race key={race.id} data={race} />;
-  }
-
-  render() {
-    return (
-      <div className="row races">
-        <div className="col-md-12">
-          <ol className="list-unstyled">
-            {this.state.data.map(this.renderRace)}
-          </ol>
-        </div>
-      </div>
-    );
-  }
-}
-
-Races.propTypes = {
-  url: React.PropTypes.string.isRequired
-};
-
-class Race extends React.Component {
   constructor(props) {
     super(props);
     this.state = {data: props.data};
@@ -89,7 +57,3 @@ class Race extends React.Component {
     );
   }
 }
-
-Race.propTypes = {
-  data: React.PropTypes.object.isRequired
-};
