+++ client/app/bundles/registration.jsx
@@ -1,6 +1,8 @@
 import ReactOnRails from 'react-on-rails';
+import Races from './components/Races';
 import RaceDetail from './components/RaceDetail';
 
 ReactOnRails.register({
+  Races,
   RaceDetail,
 });
